# perf: Fix quadratic behavior when casting Enums