--- a/crates/polars-core/src/chunked_array/logical/categorical/mod.rs
+++ b/crates/polars-core/src/chunked_array/logical/categorical/mod.rs
@@ -149,11 +149,19 @@ impl CategoricalChunked {
         };
         // Make a mapping from old idx to new idx
         let old_rev_map = self.get_rev_map();
+
+        // Create map of old category -> idx for fast lookup.
+        let old_categories = old_rev_map.get_categories();
+        let old_idx_map: PlHashMap<&str, u32> = old_categories
+            .values_iter()
+            .zip(0..old_categories.len() as u32)
+            .collect();
+
         #[allow(clippy::unnecessary_cast)]
         let idx_map: PlHashMap<u32, u32> = categories
             .values_iter()
             .enumerate_idx()
-            .filter_map(|(new_idx, s)| old_rev_map.find(s).map(|old_idx| (old_idx, new_idx as u32)))
+            .filter_map(|(new_idx, s)| old_idx_map.get(s).map(|old_idx| (*old_idx, new_idx as u32)))
             .collect();
 
         // Loop over the physicals and try get new idx